push

+++ 010_denoise_feature_dataframes.py
@@ -189,8 +189,10 @@ def construct_features_dataframe(derivatives_dir):
     # break down sca_tourettome_resid to patient and control dataframes
     sca_patients_resid = sca_tourettome_resid.drop(controls, axis=1)
     sca_controls_resid = sca_tourettome_resid.drop(patients, axis=1)
-    sca_patients_resid.to_csv(os.path.join(features_dir,'sca_patients_resid.csv'))
-    sca_controls_resid.to_csv(os.path.join(features_dir,'sca_controls_resid.csv'))
+
+    if not os.pat.isfile(os.path.join(features_dir,'sca_patients_resid.csv')):
+        sca_patients_resid.to_csv(os.path.join(features_dir,'sca_patients_resid.csv'))
+        sca_controls_resid.to_csv(os.path.join(features_dir,'sca_controls_resid.csv'))
 
     # plot sca residuals
     f = plt.figure(figsize=(17.5, 10))
@@ -232,7 +234,22 @@ def construct_features_dataframe(derivatives_dir):
         f = plt.figure(figsize=(35, 20))
         sns.heatmap(sca_controls_resid_z, yticklabels=False, cmap=cmap_gradient, vmin=-3, vmax=3)
         plt.xticks(size=6, rotation=90, weight='bold')
-        f.savefig(os.path.join(features_dir, 'sca_patients_resid_z.png'), dpi=300)
+        f.savefig(os.path.join(features_dir, 'sca_controls_resid_z.png'), dpi=300)
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
